Fix: add tarpaulin.toml and additional test coverage
In order to ensure our CI step for determining test coverage succeeds,
this commit adds a tarpaulin.toml file that configures how we output the
coverage information for CI to consume. Additionally, this commit
refactors the internals of all network related functions to accept the
api url as an argument. This allows us to stub/mock as necessary in
tests and still exercise our code without making real network
connections to the outside world.