Fix: remove dependency on openssl
In order to ensure this project builds on linux musl x86_64, this commit
swaps the tls implementation of reqwest to use rustls-tls instead of the
default of relying on openssl. This ensures we can build on linux musl
as openssl is not available by default.

Additionally, it removes the link to docs.rs as this crate currently
does not have a library component.